Develop

+++ controllers/posts.js
@@ -1,8 +1,7 @@
-const { param } = require('express/lib/request');
 const Post = require('../models/post');
+const User = require('../models/user');
 const utils = require('../utils');
 const Image = require('../models/image');
-const routers = require('../routers');
 const db = require('mongoose');
 
 function create_post(req, res, next) {
@@ -10,11 +9,10 @@ function create_post(req, res, next) {
     return utils.response(req, res, 400, {error: 'Missing required fields'});
   }
   
-  if (typeof (req.body.auth_user_id) != "string" || 
-      typeof (req.body.title) != "string" ||
+  if (typeof (req.body.title) != "string" ||
       typeof (req.body.content) != "string" ||
-      Array.isArray(req.body.images) == false ||
-      Array.isArray(req.body.keywords) == false) {
+      !Array.isArray(req.body.images) ||
+      !Array.isArray(req.body.keywords)) {
         return utils.response(req, res, 400, {error: 'Invalid type of post'});
   }
      
@@ -40,9 +38,6 @@ function create_post(req, res, next) {
     return utils.response(req, res, 400, {error: 'Content is too long'});
   }
 
-  // Validate the image paths
-  // TODO
-
   // Check if there are too many keywords
   if (req.body.keywords.length > 10) {
     return utils.response(req, res, 400, {error: 'Too many keywords'});
@@ -66,6 +61,7 @@ function create_post(req, res, next) {
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
+    keywords: req.body.keywords,
     images: req.body.images,
     posterId: req.body.auth_user_id,
   });
@@ -76,12 +72,17 @@ function create_post(req, res, next) {
     }
 
     return utils.response(req, res, 201, {message: 'Post created', postId: post._id });
-
   });
 }
-//get information about the posts, gets the post id
+
 function get_post(req, res, next) {
   const postId = req.params.id;
+
+  if (!db.Types.ObjectId.isValid(postId)){
+    return res.status(400).json({
+      error: 'Wrong get Post ID'
+    });
+  }
 
   Post.findOne({ _id: postId }, (err, post) => {
     if (err) {
@@ -92,18 +93,24 @@ function get_post(req, res, next) {
       return utils.response(req, res, 404, {error: 'Post not found'});
     }
 
-    return utils.response(req, res, 200, post); //check needed
-  });
+    return utils.response(req, res, 200, post);
+  }).populate('posterId', 'user_name profile_image bio');
 }
 
 function search_post(req, res, next) {
   var searchTerm = req.query.searchterm;
+
+  if (typeof(searchTerm) != 'string'){
+    return utils.response(req, res, 400, {error: 'Wrong searchterm type'});
+  }
+  
   var searchRegex = new RegExp('.*' + searchTerm + ".*"); //searches for any string
   const skipCount = req.query.count || 0;
 
   Post.find({ $or: [
     { content: { $regex: searchRegex } }, 
-    { title: {$regex: searchRegex} }
+    { title: { $regex: searchRegex } },
+    { keywords: { $regex: searchRegex } },
   ]}, (err, posts) => {
     if (err) {
       return utils.response(req, res, 500, {error: 'Internal server error'});
@@ -149,8 +156,11 @@ function get_post_by_userId(req, res, next) {
   const userId = req.params.id;
   const skipCount = req.query.count;
   
-  var ObjectId = require('mongoose').Types.ObjectId;
-  if (ObjectId.isValid(userId) == false){
+  if (skipCount && typeof(skipCount) != 'number'){
+    return utils.response(req, res, 400, {error: 'Wrong skipCount type'});
+  }
+
+  if (!db.Types.ObjectId.isValid(userId)){
     return res.status(400).json({
       error: 'Invalid User ID'
     });
@@ -164,43 +174,117 @@ function get_post_by_userId(req, res, next) {
     }
 
     return res.status(200).json(posts)
-  }, {skip: skipCount, limit:20}).sort({postDate:-1});
+  }, {skip: skipCount, limit:20}).populate('posterId').sort({postDate:-1});
 }
 
 function delete_post(req, res, next) {
   const postId = req.params.id;
 
-  Post.deleteOne({user: req.auth_user_id, _id: postId})
-    .then(() => {
-      return res.status(200).json({
-        message: 'Post deleted'
-      });
-    })
-    .catch(err => {
-      return res.status(500).json({
-        error: 'Internal server error'
-      });
+  if (!db.Types.ObjectId.isValid(postId)){
+    return res.status(400).json({
+      error: 'Invalid Post ID'
     });
+  }
+
+  Post.findOne({ _id: postId }, (err, post) => {
+    if (err) {
+      return utils.response(req, res, 500, {error: 'Internal server error'});
+    }
+
+    if (!post) {
+      return utils.response(req, res, 404, {error: 'Post not found'});
+    }
+
+    if (post.posterId.toString() !== req.body.auth_user_id.toString()) {
+      return utils.response(req, res, 403, {error: 'Not authorized'});
+    }
+
+    post.remove((err, post) => {
+      if (err) {
+        return utils.response(req, res, 500, {error: 'Internal server error'});
+      }
+
+      return utils.response(req, res, 200, {message: 'Post deleted'});
+    });
+  });
 }
 
 function get_newest_posts(req, res, next) {
   const skipCount = req.query.count || 0;
+
+  if (typeof(skipCount) != 'number'){
+    return utils.response(req, res, 400, {error: 'Invalid skipCount type'});
+  }
+
   Post.find((err, post) => {
     if (err) {
       return utils.response(req, res, 500, utils.response(req, res, 500, {error: 'Internal server error'}));
     }
     return utils.response(req, res, 200, post);
-  }, {skip: skipCount, limit:20}).sort({postDate:-1});
+  }, {skip: skipCount, limit:20}).populate('posterId').sort({postDate:-1});
 }
 
 function edit_post(req, res, next) {
-  if (!utils.check_body_fields(req.body, ['title', 'body', 'content', 'images','keywords'])) {
+  if (!utils.check_body_fields(req.body, [ '_id', 'auth_user_id', 'title', 'content', 'images','keywords'])) {
     return utils.response(req, res, 400, {error: 'Missing required fields'});
   }
 
-  User.findOneAndUpdate({ _id : req.body.auth_user_id }, { "$set": {
-    title: req.body.title, 
-    body: req.body.body, 
+  if (!db.Types.ObjectId.isValid(req.body._id)){
+    return res.status(400).json({
+      error: 'Invalid Post ID'
+    });
+  }
+
+  if (typeof (req.body.title) != "string" ||
+      typeof (req.body.content) != "string" ||
+      !Array.isArray(req.body.images) ||
+      !Array.isArray(req.body.keywords)) {
+        return utils.response(req, res, 400, {error: 'Invalid type of edit'});
+  }
+   
+  // Need 1 to 10 images
+  if (req.body.images.length < 1 || req.body.images.length > 10) {
+    return utils.response(req, res, 400, {error: 'Invalid number of images'});
+  }
+
+  // Check type of elements in images[] array
+  for (let i = 0; i < (req.body.images).length; i++) { 
+    if (typeof((req.body.images)[i]) != 'string') {
+      return utils.response(req, res, 400, {error: 'Invalid type of images'});
+    }
+  }
+
+  // Check if title is too long
+  if (req.body.title.length > 25) {
+    return utils.response(req, res, 400, {error: 'Title is too long'});
+  }
+
+  // Check if content is too long
+  if (req.body.content.length > 2000) {
+    return utils.response(req, res, 400, {error: 'Content is too long'});
+  }
+
+  // Check if there are too many keywords
+  if (req.body.keywords.length > 10) {
+    return utils.response(req, res, 400, {error: 'Too many keywords'});
+  }
+
+  // Check type of elements in keywords[] array
+  for (let i = 0; i < (req.body.keywords).length; i++) { 
+    if (typeof((req.body.keywords)[i]) != 'string') {
+      return utils.response(req, res, 400, {error: 'Invalid type of keywords'});
+    }
+  }
+
+  // Check if keywords are valid
+  for (let keyword of req.body.keywords) {
+    if (!utils.is_valid_keyword(keyword)) {
+      return utils.response(req, res, 400, {error: 'Invalid keywords'});
+    }
+  }
+
+  Post.findOneAndUpdate({ _id : req.body._id }, { "$set": {
+    title: req.body.title,  
     content: req.body.content,
     images: req.body.images,
     keywords: req.body.keywords
@@ -212,8 +296,15 @@ function edit_post(req, res, next) {
     }
  });
 }
+
 function delete_post_image(req, res, next){
   const imageId = req.params.id;
+
+  if (db.Types.ObjectId.isValid(imageId) == false){
+    return res.status(400).json({
+      error: 'Invalid Image ID'
+    });
+  }
 
   Image.deleteOne({_id: imageId})
     .then(() => {
